requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.3
python-dotenv==1.0.1
google-generativeai==0.8.2
openai==1.46.1
loguru==0.7.3
python-multipart==0.0.20